requirements.txt
certifi
charset-normalizer
idna
mysql-connector-python
requests
urllib3
